+++ src/app-update-checker.tsx
@@ -4,6 +4,7 @@ import styled from '@emotion/styled'
 import { fetchJSON } from 'helpers/fetch'
 import { Button } from 'components/core'
 import { useToast } from 'components/core/toast'
+import { ExternalLinkIcon } from 'components/core/icons'
 
 type Meta = {
   date: string
@@ -110,8 +111,15 @@ const AppUpdateToastContent = ({ close, version }) => {
         Application Update <span className="text-muted">{version}</span>
       </h3>
       <p style={{ marginBottom: '1.5rem' }}>
-        A new version of <i>Best of JavaScript</i> is available, click on
-        "Update" to reload the window.
+        A new version of <i>Best of JS</i> is available (
+        <a
+          href="https://github.com/bestofjs/bestofjs-webui/blob/master/CHANGELOG.md"
+          style={{ display: 'inline-flex', alignItems: 'center' }}
+        >
+          details
+          <ExternalLinkIcon />
+        </a>
+        ), click on "Update" to reload the window.
       </p>
 
       <ButtonGroup>