+++ src/components/search/search-box.tsx
@@ -98,6 +98,7 @@ export const SearchBox = () => {
 
 const Container = styled.div`
   background-color: var(--bestofjsOrange);
+  background: linear-gradient(135deg, #ed8518, #e75f16, #b94100);
   padding: 1rem 0;
 `